Extract floating point matcher to header file in libWaveletTests
