+++ src/Spectre.libWavelet/AlgorithmConstants.h
@@ -20,7 +20,7 @@
 
 namespace spectre::algorithm::wavelet
 {
-    constexpr unsigned int WAVELET_LEVELS = 10;
-    constexpr unsigned int WAVELET_BASIS = 4;
-    constexpr unsigned BASIS_LENGTH = (WAVELET_BASIS * 2) - 1;
+constexpr unsigned int WAVELET_LEVELS = 10;
+constexpr unsigned int WAVELET_BASIS = 4;
+constexpr unsigned BASIS_LENGTH = (WAVELET_BASIS * 2) - 1;
 }
